Clean sub-directories with top level make clean
Before this commit, calling make clean on the top-level Makefile would only
remove the module_load.sh file and would not recursively call the make clean's
of the sub-directories (mainly, bin_to_cube and cube_to_target).

Occasionally, especially when porting to a new system, it would be necessary to
make clean the sub-directories as they could be in a bad state (i.e. compiled
with a different compiler version/flavor), thus it being able to clean built
files in the sub-directories required one to manually cd into those directories
and call make clean.

This commit allows the ability for the top level make file to clean
subdirectories using a recursive $(MAKE) call.