Remove loading of modules from main Makefile
In order to increase portability, the load_modules target was completely
removed. Loading the modules in this manner prevents users from selecting
compiling and running the program with the compilers and libraries of their
choice, which may or may not be installed or accessed with the module command.

It is easier for the user to set up their environment (modules/libraries/path
variables etc.) themselves, which allows easier portability across system,
compilers, or NetCDF installations.

Of course, the Topo tool may not work as of this commit, but future commits
will use nc-config to correct access the correct linking and include flags
needed to compile.