Simply machine_settings.make by using nc-config
This rather large commit greatly simplifies the machine_settings.make file by
using the nc-config, which determines the compilers and locations for the active
NetCDF implementation.

First, it removes any references to a machine. This means, the user no longer
needs to manually specific the machine that they are on, or the paths to the
NetCDF installation directory.

Next, it sets FC, LDFLAGS and FFLAGS using the nc-config tool.

The -c flag was moved in the top level FFLAGS, as (I believe) all modern
C and Fortran compilers use for building and object (.o file).

Lastly, for the portion of the machine_settings.make that adds compiler
specific flags, it updates the ifeq statements to use findstring, as the
nc-config --fc might return the path of the compiler rather than just the
compiler name. This might pose a problem if the same string is within $FC, like
pgfortran and gfortran. There might be a better make tool for determine the
difference, but for now, lets just explicity check for pgfortran for gfortran
for simplicity.

Depending on a user's installation of NetCDF, users will need to
manually add flags that are not included in nc-config --flibs (such as
-lpnetcdf for NetCDF that is built with MPI).

Users will still need to edit the Makefile to point to their copy of
gmted2010_modis-ncube3000-stitch.nc for the intermediate_cube_sphere_file.